refactor(experiment-tag): move the config merge helper under util
mergeWithWindowConfig sat at src/config.ts, where the name reads like the
package's configuration module rather than one helper among the other util
functions it belongs with. Test moves with it to keep the mirror.
